02-docker/production/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
PyJWT==2.12.1
